MAINT compatibility with sklearn 1.8
