+++ .coveragerc
@@ -10,4 +10,3 @@ ignore_errors = True
 omit =
     */tests/*
     **/setup.py
-    **/_sklearn_compat.py
