+++ imblearn/ensemble/_bagging.py
@@ -18,7 +18,7 @@ from ..under_sampling import RandomUnderSampler
 from ..under_sampling.base import BaseUnderSampler
 from ..utils import Substitution, check_sampling_strategy, check_target_type
 from ..utils._docstring import _n_jobs_docstring, _random_state_docstring
-from ..utils._sklearn_compat import _fit_context
+from sklearn_compat.base import _fit_context
 
 
 @Substitution(
