+++ imblearn/ensemble/_weight_boosting.py
@@ -18,7 +18,8 @@ from ..under_sampling import RandomUnderSampler
 from ..under_sampling.base import BaseUnderSampler
 from ..utils import Substitution, check_target_type
 from ..utils._docstring import _random_state_docstring
-from ..utils._sklearn_compat import _fit_context, sklearn_version
+from sklearn_compat.base import _fit_context
+from sklearn_compat._sklearn_compat import sklearn_version
 from ._common import _adaboost_classifier_parameter_constraints
