+++ imblearn/ensemble/tests/test_forest.py
@@ -6,7 +6,7 @@ from sklearn.utils._testing import assert_allclose, assert_array_equal
 from sklearn.utils.fixes import parse_version
 
 from imblearn.ensemble import BalancedRandomForestClassifier
-from imblearn.utils._sklearn_compat import sklearn_version
+from sklearn_compat._sklearn_compat import sklearn_version
 
 
 @pytest.fixture
