+++ imblearn/pipeline.py
@@ -33,15 +33,12 @@ from sklearn.utils.metaestimators import available_if
 from sklearn.utils.validation import check_is_fitted, check_memory
 
 from .base import METHODS
-from .utils._sklearn_compat import (
-    _fit_context,
-    _print_elapsed_time,
-    _raise_for_params,
-    get_tags,
-    process_routing,
-    sklearn_version,
-    validate_params,
-)
+from sklearn_compat.base import _fit_context
+from sklearn_compat.utils import get_tags
+from sklearn_compat.utils._param_validation import validate_params
+from sklearn_compat.utils._user_interface import _print_elapsed_time
+from sklearn_compat._sklearn_compat import sklearn_version
+from sklearn_compat.utils.metadata_routing import _raise_for_params, process_routing
 
 __all__ = ["Pipeline", "make_pipeline"]
