+++ imblearn/tests/test_common.py
@@ -14,7 +14,7 @@ from sklearn.utils._testing import ignore_warnings
 
 from imblearn.over_sampling import RandomOverSampler
 from imblearn.under_sampling import RandomUnderSampler
-from imblearn.utils._sklearn_compat import (
+from sklearn_compat.utils.estimator_checks import (
     parametrize_with_checks as parametrize_with_checks_sklearn,
 )
 from imblearn.utils._test_common.instance_generator import (
