+++ imblearn/tests/test_pipeline.py
@@ -39,7 +39,8 @@ from imblearn.datasets import make_imbalance
 from imblearn.pipeline import Pipeline, make_pipeline
 from imblearn.under_sampling import EditedNearestNeighbours as ENN
 from imblearn.under_sampling import RandomUnderSampler
-from imblearn.utils._sklearn_compat import Tags, sklearn_version
+from sklearn_compat.utils._tags import Tags
+from sklearn_compat._sklearn_compat import sklearn_version
 from imblearn.utils.estimator_checks import check_param_validation
 
 JUNK_FOOD_DOCS = (
