+++ imblearn/under_sampling/_prototype_selection/_random_under_sampler.py
@@ -9,7 +9,7 @@ from sklearn.utils import _safe_indexing, check_random_state
 
 from ...utils import Substitution, check_target_type
 from ...utils._docstring import _random_state_docstring
-from ...utils._sklearn_compat import validate_data
+from sklearn_compat.utils.validation import validate_data
 from ...utils._validation import _check_X
 from ..base import BaseUnderSampler
