+++ imblearn/utils/_validation.py
@@ -16,8 +16,8 @@ from sklearn.neighbors import NearestNeighbors
 from sklearn.utils import column_or_1d
 from sklearn.utils.multiclass import type_of_target
 from sklearn.utils.validation import _num_samples
-
-from ..utils._sklearn_compat import _is_pandas_df, check_array
+from sklearn_compat.utils.validation import check_array
+from sklearn_compat.utils._dataframe import is_pandas_df
 
 SAMPLING_KIND = (
     "over-sampling",
@@ -639,7 +639,7 @@ def _check_X(X):
         raise ValueError(
             f"Found array with {n_samples} sample(s) while a minimum of 1 is required."
         )
-    if _is_pandas_df(X):
+    if is_pandas_df(X):
         return X
     return check_array(
         X, dtype=None, accept_sparse=["csr", "csc"], ensure_all_finite=False
