+++ imblearn/utils/tests/test_estimator_checks.py
@@ -6,7 +6,7 @@ from sklearn.utils.multiclass import check_classification_targets
 from imblearn.base import BaseSampler
 from imblearn.over_sampling.base import BaseOverSampler
 from imblearn.utils import check_target_type as target_check
-from imblearn.utils._sklearn_compat import validate_data
+from sklearn_compat.utils.validation import validate_data
 from imblearn.utils.estimator_checks import (
     check_samplers_fit,
     check_samplers_nan,
