+++ pyproject.toml
@@ -20,6 +20,7 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Topic :: Scientific/Engineering",
     "Topic :: Software Development :: Libraries",
 ]
@@ -28,6 +29,7 @@ dependencies = [
    "numpy>=1.25.2,<3",
    "scipy>=1.11.4,<2",
    "scikit-learn>=1.4.2,<2",
+   "sklearn-compat>=0.1.5,<0.2",
    "joblib>=1.2.0,<2",
    "threadpoolctl>=2.0.0,<4",
 ]
@@ -85,7 +87,7 @@ Issues = "https://github.com/scikit-learn-contrib/imbalanced-learn/issues"
 [tool.setuptools]
 packages = ["imblearn"]
 
-[tool.pixi.project]
+[tool.pixi.workspace]
 channels = ["conda-forge"]
 platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
 
@@ -93,6 +95,7 @@ platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
 numpy = ">=1.25.2,<3"
 scipy = ">=1.11.4,<2"
 scikit-learn = ">=1.4.2,<2"
+sklearn-compat = ">=0.1.5,<0.2"
 joblib = ">=1.2.0,<2"
 threadpoolctl = ">=2.0.0,<4"
 
@@ -187,6 +190,9 @@ python = "~=3.12.0"
 [tool.pixi.feature.py313.dependencies]
 python = "~=3.13.0"
 
+[tool.pixi.feature.py314.dependencies]
+python = "~=3.14.0"
+
 [tool.pixi.feature.tests.dependencies]
 packaging = ">=23.2,<25"
 pytest = ">=7.2.2,<9"
@@ -224,8 +230,8 @@ ci-py312-sklearn-1-6 = ["py312", "sklearn-1-6", "scipy-1-15", "tests"]
 ci-py311-latest-tensorflow = ["py311", "tensorflow", "tests"]
 ci-py311-latest-keras = ["py311", "keras", "tests"]
 
-ci-py313-latest-dependencies = ["py313", "tests"]
-ci-py313-latest-optional-dependencies = ["py313", "optional", "tests"]
+ci-py314-latest-dependencies = ["py314", "tests"]
+ci-py314-latest-optional-dependencies = ["py314", "optional", "tests"]
 
 [tool.black]
 line-length = 88
